Track each node's unmodified depth (longest chain of predecessors) as well as its priority.
Add an option to include each node's depth in the output, in addition to or instead of its priority.

Add an option to sort by depth instead of priority; this is effectively equivalent to tsort(1), although the result may be slightly different since the underlying algorithms are not the same.