Reimplement aa_destroy() in linear time.
Our current implementation of aa_destroy() performs roughly O(n*log(n)) comparisons, because it works by repeatedly removing the root element, rebalancing the tree each time.  This is completely pointless.  Replace it with a version that simply recurses to the bottom of the tree and frees nodes on the way back up.