Use a stable sorting algorithm.
The documentation promises that the output is sorted first by either priority or depth, then by either depth or priority, and finally by name.  This derives from the fact that we store nodes in a name-indexed balanced search tree in the input phase.  However, in the output phase, after calculating the depths and priorities of all nodes, we sort them using quicksort, which is not stable, breaking our promise.  Moreover, different implementations of quicksort will choose different pivots, resulting in different outcomes for equally-ranked nodes.  This is illustrated by the `t_dp_red` test case, which passed on FreeBSD but failed on Ubuntu due to different internal ordering within each priority group.

To address this, replace `qsort(3)` with `mergesort(3)` if it is available, and our own implementation of insertion sort otherwise.  Also correct the expected output of the `t_dp_red` test case.